packetdrill: tcp options in hexadecimal support
The current tcp option supports only standard tcp options,
the kind of an unknown TCP option returns an error.

We needed to test the custom tcp options, so we introduce a
hexadecimal lexicology beginning with "hex" followed by
hexadecimal data, including kind and length.

Such as
  < hex 420601020304 >
means
  tcp option { kind = 0x42, opsize = 0x06, data = 0x01020304 }

Now we can use this in our scripts:
  +0 > S 0:0(0) <mss 1460,sackOK,TS val 0 ecr 0,nop,wscale 8,nop,nop,hex 420601020304>

Signed-off-by: decly <dec.ly@163.com>