packetdrill: string ellipsis to represent any character support for getsockopt
We can use '\.hh' to represent any character in the string
argument of getsockopt, where 'hh' is ellipsis magic. This
is useful for getting some uncertain values, such as tcp's
rtt, rto, etc.

we need to make sure magic doesn't conflict with any data in
the string, and that no more than one magic can be specified
in a string.

In particular, '\...'
(a) means that magic is 0xee, or
(b) same as the previous magic

e.g.
  +0 getsockopt(4, SOL_IP, ANY_OPT, "\x01\x02\x03\x04\x05\x06\x07\x08", [8]) = 0

  Can be matched by the following:
  // using magic 0xff match '\x02' and '\x05'
  +0 getsockopt(4, SOL_IP, ANY_OPT, "\x01\.ff\x03\x04\.ff\x06\x07\x08", [8]) = 0
  // using magic 0xee (a)
  +0 getsockopt(4, SOL_IP, ANY_OPT, "\x01\...\x03\x04\...\x06\x07\x08", [8]) = 0
  // using magic 0xff, and '\...' means the previous magic 0xff (b)
  +0 getsockopt(4, SOL_IP, ANY_OPT, "\x01\.ff\x03\x04\...\x06\x07\x08", [8]) = 0

Signed-off-by: decly <dec.ly@163.com>