packetdrill: specifies the port of ICMP echoed packets support
Currently, the port of the TCP/UDP header in ICMP echoed packet
is get from the tuple of socket. And we want to be able to specify
the port of ICMP echoed packets.

e.g.
  +0 < icmp 10010>12345 unreachable frag_needed mtu 840 [1:1001(1000)]

means that the source and destination ports in the original TCP packet
of ICMP echoed are 10010 and 12345

Signed-off-by: decly <dec.ly@163.com>